refactor(combo): extract validateResponseQuality to combo/validateQuality.ts
